+++ include/GafferScene/Attributes.h
@@ -38,7 +38,7 @@
 #ifndef GAFFERSCENE_ATTRIBUTES_H
 #define GAFFERSCENE_ATTRIBUTES_H
 
-#include "GafferScene/SceneElementProcessor.h"
+#include "GafferScene/AttributeProcessor.h"
 
 #include "Gaffer/CompoundDataPlug.h"
 #include "Gaffer/TypedObjectPlug.h"
@@ -46,7 +46,7 @@
 namespace GafferScene
 {
 
-class GAFFERSCENE_API Attributes : public SceneElementProcessor
+class GAFFERSCENE_API Attributes : public AttributeProcessor
 {
 
 	public :
@@ -54,7 +54,7 @@ class GAFFERSCENE_API Attributes : public SceneElementProcessor
 		Attributes( const std::string &name=defaultName<Attributes>() );
 		~Attributes() override;
 
-		GAFFER_GRAPHCOMPONENT_DECLARE_TYPE( GafferScene::Attributes, AttributesTypeId, SceneElementProcessor );
+		GAFFER_GRAPHCOMPONENT_DECLARE_TYPE( GafferScene::Attributes, AttributesTypeId, AttributeProcessor );
 
 		Gaffer::CompoundDataPlug *attributesPlug();
 		const Gaffer::CompoundDataPlug *attributesPlug() const;
@@ -72,9 +72,9 @@ class GAFFERSCENE_API Attributes : public SceneElementProcessor
 		void hashGlobals( const Gaffer::Context *context, const ScenePlug *parent, IECore::MurmurHash &h ) const override;
 		IECore::ConstCompoundObjectPtr computeGlobals( const Gaffer::Context *context, const ScenePlug *parent ) const override;
 
-		bool processesAttributes() const override;
+		bool affectsProcessedAttributes( const Gaffer::Plug *input ) const override;
 		void hashProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
-		IECore::ConstCompoundObjectPtr computeProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, IECore::ConstCompoundObjectPtr inputAttributes ) const override;
+		IECore::ConstCompoundObjectPtr computeProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, const IECore::CompoundObject *inputAttributes ) const override;
 
 	private :