+++ include/GafferScene/SetVisualiser.h
@@ -37,7 +37,7 @@
 #ifndef GAFFERSCENE_SETVISUALISER_H
 #define GAFFERSCENE_SETVISUALISER_H
 
-#include "GafferScene/SceneElementProcessor.h"
+#include "GafferScene/AttributeProcessor.h"
 
 #include "Gaffer/NumericPlug.h"
 
@@ -60,7 +60,7 @@ namespace GafferScene
 /// display. This allows more efficient hashing/compute without the need for
 /// any internal state management, as well as permitting informative UIs that
 /// help the user understand the resultant color mappings.
-class GAFFERSCENE_API SetVisualiser : public SceneElementProcessor
+class GAFFERSCENE_API SetVisualiser : public AttributeProcessor
 {
 
 	public :
@@ -68,7 +68,7 @@ class GAFFERSCENE_API SetVisualiser : public SceneElementProcessor
 		SetVisualiser( const std::string &name=defaultName<SetVisualiser>() );
 		~SetVisualiser() override;
 
-		GAFFER_GRAPHCOMPONENT_DECLARE_TYPE( GafferScene::SetVisualiser, SetVisualiserTypeId, SceneElementProcessor );
+		GAFFER_GRAPHCOMPONENT_DECLARE_TYPE( GafferScene::SetVisualiser, SetVisualiserTypeId, AttributeProcessor );
 
 		Gaffer::StringPlug *setsPlug();
 		const Gaffer::StringPlug *setsPlug() const;
@@ -89,9 +89,9 @@ class GAFFERSCENE_API SetVisualiser : public SceneElementProcessor
 		virtual void hash( const Gaffer::ValuePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
 		virtual void compute( Gaffer::ValuePlug *output, const Gaffer::Context *context ) const override;
 
-		bool processesAttributes() const override;
+		bool affectsProcessedAttributes( const Gaffer::Plug *input ) const override;
 		void hashProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
-		IECore::ConstCompoundObjectPtr computeProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, IECore::ConstCompoundObjectPtr inputAttributes ) const override;
+		IECore::ConstCompoundObjectPtr computeProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, const IECore::CompoundObject *inputAttributes ) const override;
 
 	private :