+++ src/GafferScene/Attributes.cpp
@@ -49,17 +49,12 @@ GAFFER_GRAPHCOMPONENT_DEFINE_TYPE( Attributes );
 size_t Attributes::g_firstPlugIndex = 0;
 
 Attributes::Attributes( const std::string &name )
-	:	SceneElementProcessor( name )
+	:	AttributeProcessor( name, PathMatcher::EveryMatch )
 {
 	storeIndexOfNextChild( g_firstPlugIndex );
 	addChild( new CompoundDataPlug( "attributes" ) );
 	addChild( new BoolPlug( "global", Plug::In, false ) );
 	addChild( new AtomicCompoundDataPlug( "extraAttributes", Plug::In, new IECore::CompoundData ) );
-
-	// Fast pass-throughs for the things we don't alter.
-	outPlug()->objectPlug()->setInput( inPlug()->objectPlug() );
-	outPlug()->transformPlug()->setInput( inPlug()->transformPlug() );
-	outPlug()->boundPlug()->setInput( inPlug()->boundPlug() );
 
 	// Connect to signals we use to manage pass-throughs for globals
 	// and attributes based on the value of globalPlug().
@@ -103,17 +98,12 @@ const Gaffer::AtomicCompoundDataPlug *Attributes::extraAttributesPlug() const
 
 void Attributes::affects( const Gaffer::Plug *input, AffectedPlugsContainer &outputs ) const
 {
-	SceneElementProcessor::affects( input, outputs );
+	AttributeProcessor::affects( input, outputs );
 
 	if( attributesPlug()->isAncestorOf( input ) || input == globalPlug() || input == extraAttributesPlug() )
 	{
 		// We can only affect a particular output if we haven't
 		// connected it as a pass-through in updateInternalConnections().
-		if( !outPlug()->attributesPlug()->getInput() )
-		{
-			outputs.push_back( outPlug()->attributesPlug() );
-		}
-
 		if( !outPlug()->globalsPlug()->getInput() )
 		{
 			outputs.push_back( outPlug()->globalsPlug() );
@@ -127,7 +117,7 @@ void Attributes::hashGlobals( const Gaffer::Context *context, const ScenePlug *p
 	if( globalPlug()->getValue() )
 	{
 		// We will modify the globals.
-		SceneElementProcessor::hashGlobals( context, parent, h );
+		AttributeProcessor::hashGlobals( context, parent, h );
 		inPlug()->globalsPlug()->hash( h );
 		attributesPlug()->hash( h );
 		extraAttributesPlug()->hash( h );
@@ -175,20 +165,37 @@ IECore::ConstCompoundObjectPtr Attributes::computeGlobals( const Gaffer::Context
 	return result;
 }
 
-bool Attributes::processesAttributes() const
+bool Attributes::affectsProcessedAttributes( const Gaffer::Plug *input ) const
 {
-	// Although the base class says that we should return a constant, it should
-	// be OK to return this because it's constant across the hierarchy.
-	return ( attributesPlug()->children().size() || !extraAttributesPlug()->isSetToDefault() ) && !globalPlug()->getValue();
+	if( outPlug()->attributesPlug()->getInput() )
+	{
+		// We've made a pass-through connection
+		return false;
+	}
+
+	return
+		AttributeProcessor::affectsProcessedAttributes( input ) ||
+		attributesPlug()->isAncestorOf( input ) ||
+		input == globalPlug() ||
+		input == extraAttributesPlug()
+	;
 }
 
 void Attributes::hashProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const
 {
-	attributesPlug()->hash( h );
-	extraAttributesPlug()->hash( h );
+	if( ( attributesPlug()->children().empty() && extraAttributesPlug()->isSetToDefault() ) || globalPlug()->getValue() )
+	{
+		h = inPlug()->attributesPlug()->hash();
+	}
+	else
+	{
+		AttributeProcessor::hashProcessedAttributes( path, context, h );
+		attributesPlug()->hash( h );
+		extraAttributesPlug()->hash( h );
+	}
 }
 
-IECore::ConstCompoundObjectPtr Attributes::computeProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, IECore::ConstCompoundObjectPtr inputAttributes ) const
+IECore::ConstCompoundObjectPtr Attributes::computeProcessedAttributes( const ScenePath &path, const Gaffer::Context *context, const IECore::CompoundObject *inputAttributes ) const
 {
 	const CompoundDataPlug *ap = attributesPlug();
 	IECore::ConstCompoundDataPtr extraAttributesData = extraAttributesPlug()->getValue();