+++ src/GafferSceneModule/GlobalsBinding.cpp
@@ -43,7 +43,6 @@
 #include "GafferScene/GlobalShader.h"
 #include "GafferScene/Outputs.h"
 #include "GafferScene/Set.h"
-#include "GafferScene/SetVisualiser.h"
 
 #include "GafferBindings/DependencyNodeBinding.h"
 
@@ -112,7 +111,6 @@ void GafferSceneModule::bindGlobals()
 			;
 	}
 
-	DependencyNodeClass<SetVisualiser>();
 	DependencyNodeClass<GlobalShader>();
 
 }