+++ src/GafferSceneModule/TransformBinding.cpp
@@ -44,7 +44,6 @@
 #include "GafferScene/ParentConstraint.h"
 #include "GafferScene/PointConstraint.h"
 #include "GafferScene/Transform.h"
-#include "GafferScene/CollectTransforms.h"
 
 #include "GafferBindings/ComputeNodeBinding.h"
 
@@ -104,7 +103,5 @@ void GafferSceneModule::bindTransform()
 			.value( "ResetWorld", Transform::ResetWorld )
 		;
 	}
-
-	GafferBindings::DependencyNodeClass<GafferScene::CollectTransforms>();
 
 }